requirements.txt
asgiref==3.8.1
attrs==23.2.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
Django==4.2.11
django-bKash==0.3
django-ckeditor==6.7.1
django-colorfield==0.11.0
django-cors-headers==4.3.1
django-des==2.4.1
django-filter==24.1
django-jazzmin==3.0.0
django-js-asset==2.2.0
django-monthfield==0.1.4
django-multiselectfield==0.1.12
django-phonenumber-field==8.0.0
django-seed==0.3.1
django-solo==1.1.3
django-tinymce==4.1.0
django-wkhtmltopdf==3.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.1
et-xmlfile==1.1.0
Faker==24.9.0
idna==3.8
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
marshmallow==3.22.0
openpyxl==3.1.2
packaging==24.1
pdfkit==1.0.0
phonenumbers==8.13.44
pillow==10.3.0
psycopg-binary==3.2.1
PyJWT==2.8.0
python-dateutil==2.9.0.post0
PyYAML==6.0.1
referencing==0.34.0
reportlab==4.2.0
requests==2.32.3
rpds-py==0.18.0
six==1.16.0
sqlparse==0.4.4
sslcommerz-lib==1.0
sslcommerz-sdk==1.0.4
toposort==1.10
typed-ast==1.5.5
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
